% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pRattle.R
\name{get_scores}
\alias{get_scores}
\title{Get Document Scores from the Prattle API}
\usage{
get_scores(bank, startdate = 19980101, enddate = NA, type = "xts",
  auth.token = "Bearer bRtn76o7'g3U66?yo3S~K6(vfqLC~8", agg.level = "raw",
  dev = F)
}
\arguments{
\item{bank}{A string vector that matches a central bank shorthand code.}

\item{startdate}{Date to start the resulting data set from.}

\item{enddate}{Final date of data for data set.}

\item{type}{Defaults to an xts data frame.}

\item{auth.token}{Oauth2 token used to retreive information from Prattle API.}

\item{agg.level}{Level of aggregation for the data. Default is raw scores. Can be daily or weekly.}

\item{dev}{Pull from dev or prod?}
}
\value{
Dataframe. By default, xts/zoo type.
}
\description{
Primary R function for gathering document scores.
}
\examples{
head(get_scores('frc'))
}

