% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pRattle.R
\name{get_scores}
\alias{get_scores}
\title{Get Document Scores from the Prattle API}
\usage{
get_scores(bank, type = "xts", auth.token = NULL, agg.level = "raw")
}
\arguments{
\item{bank}{A string vector that matches a central bank shorthand code.}

\item{type}{Defaults to an xts data frame.}

\item{auth.token}{Oauth2 token (a string) used to retreive information from Prattle API.}

\item{agg.level}{Level of aggregation for the data. Default is raw scores. Can be "daily" or "weekly".}
}
\value{
Dataframe. By default, xts/zoo type.
}
\description{
Primary R function for gathering document scores.
}
\examples{
head(get_scores('frc'))
}

