% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pRattle.R
\name{get_scores}
\alias{get_scores}
\title{Get Document Scores from the Prattle API}
\usage{
get_scores(bank, type = "xts", email = NULL, pwd = NULL,
  agg.level = "raw", add.extra = TRUE)
}
\arguments{
\item{bank}{: A string vector that matches a central bank shorthand code.}

\item{type}{: Defaults to an xts data frame.}

\item{email}{: A string vector with an email for portal login}

\item{agg.level}{: Level of aggregation for the data. Default is raw scores. Can be "daily" or "weekly".}

\item{portal_pwd}{: A string vector with a password for portal login}
}
\value{
Dataframe. By default, xts/zoo type.
}
\description{
Primary R function for gathering document scores.
}
\examples{
head(get_scores('frc', email = '<YOUR EMAIL>', pwd = '<YOUR PASSWORD>'))
}

